+++ src/versions/v2/models/get-product-response-data.ts
@@ -18,9 +18,6 @@
 import { BaseProduct } from './base-product';
 // May contain unused imports in some cases
 // @ts-ignore
-import { OrganizationItemAddress } from './organization-item-address';
-// May contain unused imports in some cases
-// @ts-ignore
 import { PricesArray } from './prices-array';
 
 /**